# catalog: Avoid clones where possible