@jkosh44 2024-08-14
I did some local timings of the various slow catalog related startup sections and got the following:

## `main`

```
catalog init:           1.077803271s,   66.145475ms,    98.593723ms
durable catalog open:   96.450687ms,    91.054152ms,    86.82751ms
open catalog:           9.451893699s,   8.986238331s,   9.074442885s
```

## This PR

```
catalog init:           107.467432ms,   102.740361ms,   64.315825ms
durable catalog open:   285.891382ms,   89.369212ms,    81.850912ms
open catalog:           10.703774926s,  10.493978014s,  9.427781421s
```

It looks like there's no difference, but also these numbers are orders of magnitude smaller than what we see in staging (except for open catalog which is weirdly slower):

```
catalog init:           13.595396853s
durable catalog open:   5.28961681s
open catalog:           2.780077892s
```

so I think the local numbers are probably meaningless.

I want to do some testing in staging, but I also want to get this into the next release. It's possible that this makes no difference, but it would be very surprising if it made things worse.