@jkosh44 2024-08-14
> LGTM. I thought about the fencing logic and it made sense, but I'm unfamiliar with this code, so if you have any doubts about that code, it may be worth talking me through it in a bit more detail.

I realized I could remove this variant and replace it with an assert, so I just pushed that change which should simplify things a bit.