@jkosh44 2024-08-14
Can someone on @MaterializeInc/testing let me know if the Nightly failures are new or known: https://buildkite.com/materialize/nightly/builds/9097. They look scary but unrelated.